+++ src/polyfill/index.js
@@ -2,8 +2,8 @@
 import "core-js/proposals/global-this";
 
 // es10 string trim functions
-import "core-js/es/string/trim-left";
-import "core-js/es/string/trim-right";
+import "core-js/es/string/trim-start";
+import "core-js/es/string/trim-end";
 
 // "built-in" polyfills
 import "./console.js";