+++ src/text/bitmaptext.js
@@ -253,7 +253,7 @@ class BitmapText extends Renderable {
 
         for (var i = 0; i < this._text.length; i++) {
             x = lX;
-            var string = this._text[i].trimRight();
+            var string = this._text[i].trimEnd();
             // adjust x pos based on alignment value
             var stringWidth = this.metrics.lineWidth(string);
             switch (this.textAlign) {