+++ CHANGELOG.md
@@ -11,8 +11,9 @@
 - WebGL: dissociate the `preMultipliedAlpha` setting from the `transparent` one since those are 2 different things
 
 ### Fixed
-- WebGL: fix a regression with global opacity not being properly cascaded to texture in WebGL mode (thanks @wpernath)
+- BitmapText: fix improper `fillStyle` initialization in BitmapText (thanks @wpernath)
 - Canvas: fix the "transparent" canvas mode with the canvas renderer (thanks @wpernath)
+- WebGL: fix a regression with global opacity not being properly cascaded to texture in WebGL mode (thanks @wpernath)
 
 ## [13.1.1] (melonJS 2) - _2022-08-10_