+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## [13.2.0] (melonJS 2) - _2022-xx-xx_
+## [13.2.1] (melonJS 2) - _2022-08-22_
+
+### Fixed
+- Release date and version/change tracking
+
+## [13.2.0] (melonJS 2) - _2022-08-22_
 
 ### Added
 - Geometry: new base Point geometry object
@@ -1298,6 +1303,7 @@
 - first public version with alex4 !
 - www.melonjs.org is live :)
 
+[13.2.1]: https://github.com/melonjs/melonJS/compare/13.2.1...13.2.0
 [13.2.0]: https://github.com/melonjs/melonJS/compare/13.2.0...13.1.1
 [13.1.1]: https://github.com/melonjs/melonJS/compare/13.1.1...13.1.0
 [13.1.0]: https://github.com/melonjs/melonJS/compare/13.1.0...13.0.0